Multiple changes:
Changes to folder structure, now we stronlgy align to Maven design
Changes to the way credentials are setup-> all credentials in
Config.json only
Updated maven dependency version to 1.0.2 for TestZeus.
Fixed the bug regarding "testzeus.base" not found in Basetest class.
🚀
