+++ .gitignore
@@ -30,3 +30,4 @@ target/
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+/desktop.ini
