+++ pom.xml
@@ -13,7 +13,7 @@
 		<selinium.java.version>3.141.59</selinium.java.version>
 		<apache.httpcomponents.core.version>4.4.4</apache.httpcomponents.core.version>
 		<apache.httpcomponents.client.version>4.5.2</apache.httpcomponents.client.version>
-		<testng.version>7.0.0</testng.version>
+		<testng.version>7.6.0</testng.version>
 		<commons-lang.version>2.6</commons-lang.version>
 		<super-csv.version>2.4.0</super-csv.version>
 		<jsoup.version>1.10.1</jsoup.version>
@@ -31,7 +31,7 @@
 	<dependency>
   <groupId>com.testzeus</groupId>
   <artifactId>Test_Zeus</artifactId>
-  <version>1.0.1</version>
+  <version>1.0.2</version>
 </dependency>
 
 		<!-- https://mvnrepository.com/artifact/com.google.guava/guava -->
@@ -51,18 +51,17 @@
 			<version>1.3</version>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-			<groupId>com.google.code.gson</groupId>
-			<artifactId>gson</artifactId>
-			<version>2.8.2</version>
-			<scope>test</scope>
-		</dependency>
+		
 		<dependency>
 			<groupId>org.seleniumhq.selenium</groupId>
 			<artifactId>selenium-java</artifactId>
 			<version>${selinium.java.version}</version>
 		</dependency>
-
+<dependency>
+			<groupId>com.google.code.gson</groupId>
+			<artifactId>gson</artifactId>
+			<version>2.8.2</version>
+		</dependency>
 		<dependency>
 			<groupId>org.apache.httpcomponents</groupId>
 			<artifactId>httpcore</artifactId>
@@ -177,8 +176,7 @@
 <dependency>
     <groupId>io.github.bonigarcia</groupId>
     <artifactId>webdrivermanager</artifactId>
-    <version>5.0.3</version>
-    <scope>test</scope>
+    <version>5.3.0</version>
 </dependency>
 
   <dependency>
@@ -191,6 +189,11 @@
     <artifactId>log4j-core</artifactId>
     <version>2.16.0</version>
   </dependency> 
+  <dependency>
+    <groupId>ai.dev-tools</groupId>
+    <artifactId>ai-devtools-selenium</artifactId>
+    <version>0.1.6</version>
+</dependency>
 	</dependencies>
 
 	<profiles>
