+++ src/main/java/base/Log.java
@@ -1,4 +1,4 @@
-package com.AT.base;
+package base;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
