+++ src/main/java/base/PropertyReader.java
@@ -1,4 +1,4 @@
-package com.AT.base;
+package base;
 
 import java.util.Properties;
