+++ src/main/java/pageobjects/.gitignore
@@ -1,0 +1,1 @@
+/desktop.ini
