+++ src/main/java/pageobjects/AccountListPage.java
@@ -1,4 +1,4 @@
-package com.AT.pageobjects;
+package pageobjects;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
