+++ src/test/java/testscripts/.gitignore
@@ -1,0 +1,1 @@
+/desktop.ini
