+++ src/test/resources/.gitignore
@@ -1,0 +1,1 @@
+/desktop.ini
