+++ src/test/resources/all-tests-chrome.xml
@@ -7,7 +7,9 @@
 	<test name="smoke_test" preserve-order="true">
 		<classes>
 
-			<class name="com.AT.testscripts.AccountCreationViaUI" />
+			<class name="testscripts.AccountCreationViaUI" />
+						<class name="testscripts.AccountCreationViaAPIs" />
+			
 
 		</classes>
 	</test>
